simple cli version info
