+++ Makefile
@@ -1,10 +1,11 @@
-
+MODULE   = github.com/dhouti/sops-converter
 DATE    ?= $(shell date +%FT%T%z)
 VERSION ?= $(shell git describe --tags --always --dirty --match=v* 2> /dev/null || \
 			cat $(CURDIR)/.version 2> /dev/null || echo v0)
 GIT_COMMIT_ID ?= $(shell git rev-parse --short HEAD)
 
 LDFLAGS = "-X main.AppVersion=$(VERSION) -X main.BuildDate=$(DATE) -X main.GitCommit=$(GIT_COMMIT_ID)"
+CLI_LDFLAGS = "-X $(MODULE)/cli/cmd.AppVersion=$(VERSION) -X $(MODULE)/cli/cmd.BuildDate=$(DATE) -X $(MODULE)/cli/cmd.GitCommit=$(GIT_COMMIT_ID)"
 
 # Image URL to use all building/pushing image targets
 IMG ?= docker.io/$(DOCKER_USER)/sops-converter:v0.2.2
@@ -21,16 +22,16 @@ endif
 all: fmt vet generate manifests mocks
 
 build: generate manifests fmt vet
-	go build -ldflags $(LDFLAGS) -o bin/sops-converter github.com/dhouti/sops-converter/cli
+	go build -ldflags $(CLI_LDFLAGS) -o bin/sops-converter github.com/dhouti/sops-converter/cli
 	go build -ldflags $(LDFLAGS) -o bin/controller github.com/dhouti/sops-converter
 
 release-cli: build-cli zip-cli
 
 build-cli: generate manifests fmt vet
-	GOOS=darwin GOARCH=amd64 go build -ldflags $(LDFLAGS) -o bin/sops-converter-cli-darwin-amd64 github.com/dhouti/sops-converter/cli
-	GOOS=linux GOARCH=amd64 go build -ldflags $(LDFLAGS) -o bin/sops-converter-cli-linux-amd64 github.com/dhouti/sops-converter/cli
-	GOOS=linux GOARCH=arm64 go build -ldflags $(LDFLAGS) -o bin/sops-converter-cli-linux-arm64 github.com/dhouti/sops-converter/cli
-	GOOS=windows GOARCH=amd64 go build -ldflags $(LDFLAGS) -o bin/sops-converter-cli-windows-amd64 github.com/dhouti/sops-converter/cli
+	GOOS=darwin GOARCH=amd64 go build -ldflags $(CLI_LDFLAGS) -o bin/sops-converter-cli-darwin-amd64 github.com/dhouti/sops-converter/cli
+	GOOS=linux GOARCH=amd64 go build -ldflags $(CLI_LDFLAGS) -o bin/sops-converter-cli-linux-amd64 github.com/dhouti/sops-converter/cli
+	GOOS=linux GOARCH=arm64 go build -ldflags $(CLI_LDFLAGS) -o bin/sops-converter-cli-linux-arm64 github.com/dhouti/sops-converter/cli
+	GOOS=windows GOARCH=amd64 go build -ldflags $(CLI_LDFLAGS) -o bin/sops-converter-cli-windows-amd64 github.com/dhouti/sops-converter/cli
 
 zip-cli:
 	zip -rq bin/sops-converter-cli-windows-amd64.zip bin/sops-converter-cli-windows-amd64
