+++ cli/cmd/version.go
@@ -1,0 +1,33 @@
+package cmd
+
+import (
+	"fmt"
+	log "github.com/sirupsen/logrus"
+	"github.com/spf13/cobra"
+	goruntime "runtime"
+)
+
+var (
+	AppVersion, BuildDate, GitCommit string
+)
+
+func init() {
+	rootCmd.AddCommand(versionCmd)
+}
+
+var versionCmd = &cobra.Command{
+	Use:   "version",
+	Short: "Print version",
+	Long:  `Display version and build information about rapidkube.`,
+	Run: func(cmd *cobra.Command, args []string) {
+		printVersion()
+	},
+}
+
+func printVersion() {
+	log.Info(fmt.Sprintf("Version: %s", AppVersion))
+	log.Info(fmt.Sprintf("Go Version: %s", goruntime.Version()))
+	log.Info(fmt.Sprintf("Go OS/Arch: %s/%s", goruntime.GOOS, goruntime.GOARCH))
+	log.Info(fmt.Sprintf("Git Commit: %s", GitCommit))
+	log.Info(fmt.Sprintf("BuildDate: %s", BuildDate))
+}
