+++ cli/main.go
@@ -17,30 +17,14 @@ limitations under the License.
 package main
 
 import (
-	"fmt"
 	"github.com/dhouti/sops-converter/cli/cmd"
 	"github.com/dhouti/sops-converter/cli/logger"
-	log "github.com/sirupsen/logrus"
-	goruntime "runtime"
-)
-
-var (
-	AppVersion, BuildDate, GitCommit string
 )
 
 func init() {
 	logger.ConfigureLogging(nil)
 }
 
-func printVersion() {
-	log.Info(fmt.Sprintf("Version: %s", AppVersion))
-	log.Info(fmt.Sprintf("Go Version: %s", goruntime.Version()))
-	log.Info(fmt.Sprintf("Go OS/Arch: %s/%s", goruntime.GOOS, goruntime.GOARCH))
-	log.Info(fmt.Sprintf("Git Commit: %s", GitCommit))
-	log.Info(fmt.Sprintf("BuildDate: %s", BuildDate))
-}
-
 func main() {
-	printVersion()
 	cmd.Execute()
 }
